wiz-backend/freelance/requirements.txt
django
djangorestframework
django-cors-headers
gunicorn
whitenoise
dj-database-url
psycopg2-binary
pillow
python-dotenv 